simple_python_projects/Antlr4TsqlParser/.travis.requirements.txt
flake8
tox
cmd2
flask
Flask-Testing
antlr4-python3-runtime
coverage
pytest
pytest-flask
pytest-cov